+++ domains/Locations/Models/Location.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Domains\Locations\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class Location extends Model
+{
+    use SoftDeletes;
+
+    protected $fillable = [
+        'customer_id',
+        'name',
+        'parent_id',
+    ];
+}