+++ domains/Locations/routes.php
@@ -1,0 +1,6 @@
+<?php
+
+use Domains\Locations\Http\Controllers\LocationsStoreController;
+use Illuminate\Support\Facades\Route;
+
+Route::post('/locations', LocationsStoreController::class);